Add edit cabin functionality

+++ src/services/apiCabins.ts
@@ -2,7 +2,9 @@ import type { Database } from "../types/supabase";
 import supabase, { supabaseUrl } from "./supabase";
 
 type Cabin = Database["public"]["Tables"]["cabins"];
-type CabinInsert = Omit<Cabin["Insert"], "image"> & { image: File };
+type CabinInsertDb = Cabin["Insert"];
+type CabinUpdateDb = Cabin["Update"];
+type CabinInsert = Omit<Cabin["Insert"], "image"> & { image: string | File };
 
 export async function getCabins() {
     const { data, error } = await supabase.from("cabins").select("*");
@@ -15,19 +17,37 @@ export async function getCabins() {
     return data;
 }
 
-export async function createCabin(newCabin: CabinInsert) {
-    const imageName = `${Math.random()}-${newCabin.image.name}`.replaceAll(
-        "/",
-        ""
-    );
-    const imagePath = `${supabaseUrl}/storage/v1/object/public/cabin-images/${imageName}`;
+export async function createEditCabin(newCabin: CabinInsert, id?: number) {
+    const hasImagePath = typeof newCabin.image === "string";
 
-    // Create Cabin
-    const { data, error } = await supabase
-        .from("cabins")
-        .insert([{ ...newCabin, image: imagePath }])
-        .select()
-        .single();
+    let imageName: string;
+    if (typeof newCabin.image === "string") {
+        imageName = newCabin.image;
+    } else {
+        imageName = `${Math.random()}-${newCabin.image.name}`.replaceAll(
+            "/",
+            ""
+        );
+    }
+
+    const imagePath = hasImagePath
+        ? newCabin.image
+        : `${supabaseUrl}/storage/v1/object/public/cabin-images/${imageName}`;
+
+    const { data, error } = await (!id
+        ? // Create
+          supabase
+              .from("cabins")
+              .insert([{ ...newCabin, image: imagePath } as CabinInsertDb])
+              .select()
+              .single()
+        : // Update
+          supabase
+              .from("cabins")
+              .update({ ...newCabin, image: imagePath } as CabinUpdateDb)
+              .eq("id", id)
+              .select()
+              .single());
 
     if (error) {
         console.error(error);
@@ -35,17 +55,19 @@ export async function createCabin(newCabin: CabinInsert) {
     }
 
     // Upload image
-    const { error: storageError } = await supabase.storage
-        .from("cabin-images")
-        .upload(imageName, newCabin.image);
+    if (!hasImagePath) {
+        const { error: storageError } = await supabase.storage
+            .from("cabin-images")
+            .upload(imageName, newCabin.image);
 
-    // Delete the cabin if there was an error uploading image
-    if (storageError) {
-        await supabase.from("cabins").delete().eq("id", data.id);
-        console.error(storageError);
-        throw new Error(
-            "Cabins image could not be uploaded and the cabin was not created"
-        );
+        // Delete the cabin if there was an error uploading image
+        if (storageError) {
+            await supabase.from("cabins").delete().eq("id", data.id);
+            console.error(storageError);
+            throw new Error(
+                "Cabins image could not be uploaded and the cabin was not created"
+            );
+        }
     }
 
     return data;
